+++ src/tests/integration/TestFiles/ProviderOptions.cs
@@ -14,9 +14,21 @@ class ProviderOptions
     [OutputFormatProvider(typeof(ProviderOptions), nameof(MyDoubleNoQuotes), true)]
     public double? DoublePropNullableNull { get; set; }
 
+    [OutputFormatProvider(typeof(ProviderOptions), nameof(MyStringQuotes))]
+    public string? StringPropNullable { get; set; } = "This is a long string";
+
+    [OutputFormatProvider(typeof(ProviderOptions), nameof(MyStringQuotes), true)]
+    public string? StringPropNullableNull { get; set; } = "This is a longer string";
+
+    public static string? MyStringQuotes(string? o)
+    {
+        if (o is null) return null;
+        return o.Replace("i","ZZZ");
+    }
+
     public static string? MyDoubleNoQuotes(double? o)
     {
         if (o is null) return null;
-        return o.Value.ToString("0.00");
+        return o.Value.ToString("0.000");
     }
 }